Use diamond storage layout (#103)
* add basic diamond storage setup

* refactor, use _REFLEX_STORAGE

* clean up

* remove sentinel

* fix snapshot script

* clean up

* clean up

* add note, reorder

* clean up

* remove payable

* add limitations and gas shortcut method

* start updating gas tests, weird batch gas issue

* add gas usage

* update gas measurements

* update note

* clear up

* use most minimal version of ReflexBatch, old version was distorting gas measurements

* always pass back violating value in revert message for easier debugging

* recreate storage report and add note

* optimize reentrancy guard for hot path

* update malicious module to explain you cannot cause storage clashes by hash re-use

* update internal module diagram

* fix previous commit

* fix previous commit

* revert optimized reentrancy status changes for baseline

* add brutalizer to addresses as preperation for gas golfing

* clean up

* clean up

* clean up

* strange issue with module tests passing in storage clash

* issue was caused by function selector clashing, should never be an actual problem

* clean up

* revert gas report creation, not very useful considering it specifies deployment cost - not runtime cost

* update stats

* clean up

* fix version

* the current storage scripts rely on using Solidity storage, for now remove - at a later point re-introduce scripts using a custom layout checker which knows about diamond storage

* clean up

* update naming

* clean up docs

* revert named mapping parameters for min-solc

+++ scripts/snapshot.sh
@@ -32,15 +32,23 @@ export FOUNDRY_PROFILE=$PROFILE
 
 log $GREEN "Creating snapshot with profile: $PROFILE"
 
+# Register if old snapshot exists
+HAS_OLD_SNAPSHOT=false
+
 # Back up old snapshot
-cp .gas-snapshot-$PROFILE .gas-snapshot-$PROFILE-old
+if [ -f .gas-snapshot-$PROFILE ]; then
+  cp .gas-snapshot-$PROFILE .gas-snapshot-$PROFILE-old
+  HAS_OLD_SNAPSHOT=true
+fi
 
 # Create snapshot
 forge snapshot \
     --match-test $SCOPE \
     --snap .gas-snapshot-$PROFILE
 
-# Generate diff
-diff .gas-snapshot-$PROFILE .gas-snapshot-$PROFILE-old > .gas-snapshot-$PROFILE-diff
+# # Generate diff
+if [ "$HAS_OLD_SNAPSHOT" = true ]; then
+  diff .gas-snapshot-$PROFILE .gas-snapshot-$PROFILE-old > .gas-snapshot-$PROFILE-diff
+fi
 
 log $GREEN "Done"
